Moved parseStringsArray from User.js data model to data.js; added documentation to data.js functions
